+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 # Changelog
 
 -----------------------------------------------
+[0.2.3] - 2021.01.08
+
+* Fix backToForeground method.
+
+[0.2.2] - 2020.12.27
+
+* Update json format for push payload.
+
+[0.2.1] - 2020.12.23
+
+* Fix: Missing null check.
+* Fix: change parameter handle to number.
+
 [0.2.0] - 2020.11.11
 
 * Change FlutterCallKeep as a singleton.